Attempt to simplify the feature model. The ISO 19109 specification said that attributes may have a [0 ... n] cardinality, but does not said how multi-occurrences should be represented. There is 2 ways to represent an occurrence of n attributes:
  1) As n instance of Attribute, where each Attribute contains one value.
  2) As only once instance of Attribute, which contains n value (e.g. as a collection).

The old GeoAPI model in the org.opengis.feature package was implementing approach 1, so our previous org.apache.sis.feature
model was following the same path.  However after seeing where this approach was leading us (more complexity in the form of
special PropertyList and PropertySingleton implementations for keeping decent performances) and after discussion with users
of previous model, we feel that approach 2 would be both simpler and a better fit for users of Features.

In this commit, we change the cardinality of 'DefaultFeature.getProperty(String)' from [0 ... n] to [0 ... 1].
A future commit will need to change the cardinality of 'DefaultAttribute.getValue()' from [0 ... 1] to [0 ... n],
with a convenience method for the very common case where we know that n = 1.


git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK8@1593149 13f79535-47bb-0310-9956-ffa450edef68